feat: 添加全局中文字体支持并优化SVD模态分析
refactor(utils): 统一移除局部字体设置，改用全局sitecustomize配置
fix(pde_flow): 修正ProcessedPDEBenchDataset中idx的数据类型为torch.long
docs: 添加多个模态分析报告和SVD训练示例代码
build: 添加SourceHanSansSC字体文件
style: 清理多个文件中的冗余字体配置代码

+++ generate_data/pde_flow/train_with_processed_data.py
@@ -268,7 +268,7 @@ class ProcessedPDEBenchDataset(torch.utils.data.Dataset):
     
     def __getitem__(self, idx):
         # 返回 (input, output, time_step)
-        return self.inputs[idx], self.outputs[idx], torch.tensor(idx, dtype=torch.float32)
+        return self.inputs[idx], self.outputs[idx], torch.tensor(idx, dtype=torch.long)
     
     def get_data_statistics(self):
         """获取数据统计信息"""
